requirements.txt
fastapi
uvicorn
pydantic
numpy
Pillow
scikit-learn
python-multipart